refactor: homepage styles moved from navbar to layout
